pyproject.toml
[tool.poetry]
name = "ethereum-watcher"
version = "0.1.0"
description = "A ethereum viewer"
authors = ["Lucho Leves <lucianoleves@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.3"
etherscan-python = "^2.1.0"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
celery = "^5.1.2"
circus = "^0.17.1"
django-celery-beat = "^2.2.1"
psycopg2-binary = "^2.9.1"
pytest-mock = "^3.6.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
pre-commit = "^2.12.0"
pdbpp = "^0.10.2"
bpython = "^0.21"
faker = "^8.0.0"
factory-boy = "^3.2.0"
pytest-django = "^4.4.0"
pytest-factoryboy = "^2.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
